Move api stuff on api application
